# docs/api_doc/swagger.json
{"openapi":"3.0.1","info":{"title":"'app'","description":"'app'","contact":{"name":"None","email":"none@none.com"},"version":"'0.0.1-SNAPSHOT'"},"servers":[{"url":"http://localhost:8081","description":"'app'"}],"paths":{"/ping":{"get":{"tags":["ping-controller"],"summary":"Check healthy of the app","description":"If the app it's alive response pong","operationId":"pong","responses":{"200":{"description":"Successful operation","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ErrorApi"}}}}}}},"/elecciones/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultadoNacional","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoNacionalDTO"}}}}}}}},"/elecciones/distritos":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritos","parameters":[{"name":"id","in":"query","required":false,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DistritoDTO"}}}}}}}},"/elecciones/distritos/{idDistrito}/secciones":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritosSinSeccion","parameters":[{"name":"idDistrito","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}}},"/elecciones/distritos/{idDistrito}/secciones/{idSeccion}":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritoConSeccion","parameters":[{"name":"idDistrito","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idSeccion","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SeccionDTO"}}}}}}}},"/elecciones/distritos/{idDistrito}/resultados":{"get":{"tags":["elecciones-controller"],"operationId":"getResultadoByDistrito","parameters":[{"name":"idDistrito","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDTO"}}}}}}}},"/elecciones/distritos/{idDistrito}/cargos":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritosSinCargo","parameters":[{"name":"idDistrito","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CargosDTO"}}}}}}}},"/elecciones/distritos/{idDistrito}/cargos/{idCargo}":{"get":{"tags":["elecciones-controller"],"operationId":"getDistritosConCargo","parameters":[{"name":"idDistrito","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"idCargo","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CargosDTO"}}}}}}}}},"components":{"schemas":{"ErrorApi":{"type":"object","properties":{"timestamp":{"type":"string"},"status":{"type":"integer","format":"int32"},"error":{"type":"string"},"message":{"type":"string"}}},"ResultadoDistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"secciones":{"type":"array","items":{"type":"string"}},"votos_escrutados":{"type":"integer","format":"int64"},"porcentaje_padron_nacional":{"type":"number","format":"double"},"agrupacion_ganadora":{"type":"string"},"resultados_agrupaciones":{"type":"array","items":{"$ref":"#/components/schemas/ResultadosAgrupacionesDTO"}}}},"ResultadoNacionalDTO":{"type":"object","properties":{"distritos":{"type":"array","items":{"type":"string"}},"votos_escrutados":{"type":"integer","format":"int64"},"agrupacion_ganadora":{"type":"string"},"resultados_nacionales":{"type":"array","items":{"$ref":"#/components/schemas/ResultadosAgrupacionesDTO"}},"resultados_distritos":{"type":"array","items":{"$ref":"#/components/schemas/ResultadoDistritoDTO"}}}},"ResultadosAgrupacionesDTO":{"type":"object","properties":{"nombre":{"type":"string"},"posicion":{"type":"integer","format":"int32"},"votos":{"type":"integer","format":"int64"},"porcentaje":{"type":"string"}}},"DistritoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"SeccionDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}},"ResultadoDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"},"votantes":{"type":"integer","format":"int64"},"secciones":{"type":"array","items":{"type":"string"}},"porcentaje_padron_nacional":{"type":"number","format":"double"},"agrupacion_ganadora":{"type":"string"},"resultados_agrupaciones":{"type":"array","items":{"$ref":"#/components/schemas/ResultadosAgrupacionesDTO"}}}},"CargosDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nombre":{"type":"string"}}}}}}