# src/main/java/ar/edu/utn/frc/tup/lciv/dtos/CargosDTO.java
package ar.edu.utn.frc.tup.lciv.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class CargosDTO {
    private Long id;
    private String nombre;
}