# src/main/java/ar/edu/utn/frc/tup/lciv/dtos/DistritoDTO.java
package ar.edu.utn.frc.tup.lciv.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class DistritoDTO {
    private Long id;
    private String nombre;
}