# src/main/java/ar/edu/utn/frc/tup/lciv/dtos/SeccionDTO.java
package ar.edu.utn.frc.tup.lciv.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@AllArgsConstructor
@NoArgsConstructor
@Data
public class SeccionDTO {
    private Long id;
    private String nombre;
}